refactor: Extracted custom chart FilterGitTree to abstract filterTree function in utils
